+++ react/index.js
@@ -1,0 +1,2 @@
+
+module.exports = require('../build/react');
