+++ src/react/index.js
@@ -1,0 +1,3 @@
+
+export { default as reducerComponent } from  './reducerComponent';
+export { default as useReducer } from  './useReducer';
