+++ src/react/reducerComponent.js
